CS-341 Rename components
