+++ src/components/shared/ClearIcon.js
